refactor(architecture): fix the remaining 4 ADR-061 debt-no-growth violations
Closes the rest of the debt-no-growth gate scripts/check_architecture.py
was reporting on main (the first two of six were fixed in the prior
commit on this branch):

- cli_buildsource_helpers.py (frontends, +2 over baseline): moved
  _purge_external_outputs -- self-contained, zero first-party deps -- to
  buildsource/pack_shape.py (extract), reached through
  workflows/extraction.py's existing re-export facade rather than a
  direct frontends -> extract import (which check_architecture.py's
  dependency-direction check correctly rejects for a module-level
  import).

- serialization.py (unclassified, "storage (blocked, deeper than
  expected)", +11 over baseline): moved
  _extraction_contract_from_dict/_build_mode_from_dict -- on-load format
  migrations, not comparison logic -- into
  storage/snapshot_load_normalization.py, which already exists
  specifically for "new load-time migration logic" this debt-baselined
  file can't grow to hold. BuildMode/its enums import from
  model/build_mode_facts.py directly (already the real definition
  location; build_mode.py only re-exports it) rather than through the
  extract-classified build_mode.py, so the storage-classified
  destination stays direction-clean.

- test_mutation_score_gate.py (+36 over baseline): split TestUngatedRun
  into tests/test_mutation_score_gate_ungated.py, matching the sibling
  split test_mutation_run_scoping.py -> test_mutation_per_module_scoping.py
  already established for this same gate script.

- test_service_unit.py (+1 over baseline): split TestResolveSidesSequentially
  into tests/test_service_compare_pipeline.py (service_compare_pipeline.py
  had no dedicated test file yet).

architecture/debt.yaml: baselines updated to match; test_mutation_score_gate.py's
true pre-fix baseline was found to be 1310 lines, not the 1300 an earlier
measurement in this session mistakenly recorded.

Found and deliberately left alone: tests/test_service_unit.py has
pre-existing `ruff format` drift unrelated to this change (confirmed
reproducing identically on origin/main) -- several `patch(...)` call
sites grew past 88 columns when two ADR-061 commits lengthened their
mocked-module-path strings without a `ruff format` pass afterward.
Reformatting that file's ~170 unrelated lines is out of scope here.

Verification: check_architecture.py 6 errors -> 0. mypy abicheck/ clean
(554 files). ruff check clean; ruff format --check clean on every file
this commit actually authored (test_service_unit.py's pre-existing,
unrelated drift excepted). 390+ targeted tests across every touched
module pass unchanged, plus the full fast unit suite.
